proyecto-frameffx/requirements.txt
asgiref==3.10.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
crispy-bootstrap5==2025.6
cryptography==46.0.3
defusedxml==0.7.1
Django==5.2
django-allauth==0.61.0
django-crispy-forms==2.5
djangorestframework==3.14.0
Faker==38.2.0
gunicorn==23.0.0
idna==3.11
oauthlib==3.3.1
packaging==25.0
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.0.0
python3-openid==3.2.0
psycopg2-binary==2.9.9
requests==2.32.5
requests-oauthlib==2.0.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.3
